Move homepage components into own folder
