+++ client/src/components/homepage/WordOfTheDay.js
